fix(monolith): DeviceTypeService -> DeviceTypeComponent
