+++ smart-home-monolith/pom.xml
@@ -105,13 +105,7 @@
                     </excludes>
                 </configuration>
             </plugin>
-            <plugin>
-                <groupId>org.springframework.boot</groupId>
-                <artifactId>spring-boot-maven-plugin</artifactId>
-                <configuration>
-                    <mainClass>com.baeldung.spring.kafka.KafkaApplication</mainClass>
-                </configuration>
-            </plugin>
+            
         </plugins>
     </build>
 </project>
