+++ smart-home-monolith/src/main/java/ru/yandex/practicum/smarthome/service/HeatingSystemService.java
@@ -1,13 +1,76 @@
 package ru.yandex.practicum.smarthome.service;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 import ru.yandex.practicum.smarthome.dto.HeatingSystemDto;
+import ru.yandex.practicum.smarthome.dto.CommandDto;
 import ru.yandex.practicum.smarthome.entity.HeatingSystem;
+import ru.yandex.practicum.smarthome.repository.HeatingSystemRepository;
+import ru.yandex.practicum.smarthome.components.KafkaProducer;
+import ru.yandex.practicum.smarthome.components.DeviceType;
 
-public interface HeatingSystemService {
-    HeatingSystemDto getHeatingSystem(Long id);
-    HeatingSystemDto updateHeatingSystem(Long id, HeatingSystemDto heatingSystemDto);
-    void turnOn(Long id);
-    void turnOff(Long id);
-    void setTargetTemperature(Long id, double temperature);
-    Double getCurrentTemperature(Long id);
+
+@Service
+@RequiredArgsConstructor
+public class HeatingSystemService {
+    private final HeatingSystemRepository heatingSystemRepository;
+    private final KafkaProducer kafkaProducer;
+    private final DeviceType deviceType;
+    
+    public HeatingSystemDto getHeatingSystem(Long id) {
+        HeatingSystem heatingSystem = heatingSystemRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("HeatingSystem not found"));
+        return convertToDto(heatingSystem);
+    }
+
+    public HeatingSystemDto updateHeatingSystem(Long id, HeatingSystemDto heatingSystemDto) {
+        HeatingSystem existingHeatingSystem = heatingSystemRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("HeatingSystem not found"));
+        existingHeatingSystem.setOn(heatingSystemDto.isOn());
+        existingHeatingSystem.setTargetTemperature(heatingSystemDto.getTargetTemperature());
+        HeatingSystem updatedHeatingSystem = heatingSystemRepository.save(existingHeatingSystem);
+        return convertToDto(updatedHeatingSystem);
+    }
+
+    public void turnOn(Long id) {
+        HeatingSystem heatingSystem = heatingSystemRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("HeatingSystem not found"));
+        heatingSystem.setOn(true);
+        heatingSystemRepository.save(heatingSystem);
+
+        this.kafkaProducer.sendMessage(this.deviceType.GetHeatingSystemOnCommand());
+    }
+
+    public void turnOff(Long id) {
+        HeatingSystem heatingSystem = heatingSystemRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("HeatingSystem not found"));
+        heatingSystem.setOn(false);
+        heatingSystemRepository.save(heatingSystem);
+
+        this.kafkaProducer.sendMessage(this.deviceType.GetHeatingSystemOffCommand());
+    }
+
+    public void setTargetTemperature(Long id, double temperature) {
+        HeatingSystem heatingSystem = heatingSystemRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("HeatingSystem not found"));
+        heatingSystem.setTargetTemperature(temperature);
+        heatingSystemRepository.save(heatingSystem);
+
+        this.kafkaProducer.sendMessage(this.deviceType.GetHeatingSystemSetCommand(temperature));
+    }
+
+    public Double getCurrentTemperature(Long id) {
+        // TODO: get from telemetry service
+        HeatingSystem heatingSystem = heatingSystemRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("HeatingSystem not found"));
+        return heatingSystem.getCurrentTemperature();
+    }
+
+    private HeatingSystemDto convertToDto(HeatingSystem heatingSystem) {
+        HeatingSystemDto dto = new HeatingSystemDto();
+        dto.setId(heatingSystem.getId());
+        dto.setOn(heatingSystem.isOn());
+        dto.setTargetTemperature(heatingSystem.getTargetTemperature());
+        return dto;
+    }
 }
