--- smart-home-monolith/src/main/java/ru/yandex/practicum/smarthome/service/HeatingSystemServiceImpl.java
@@ -1,89 +1,0 @@
-package ru.yandex.practicum.smarthome.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.kafka.core.KafkaTemplate;
-import ru.yandex.practicum.smarthome.dto.HeatingSystemDto;
-import ru.yandex.practicum.smarthome.dto.CommandDto;
-import ru.yandex.practicum.smarthome.entity.HeatingSystem;
-import ru.yandex.practicum.smarthome.repository.HeatingSystemRepository;
-import ru.yandex.practicum.smarthome.kafka.KafkaProducerConfig;
-
-@Service
-@RequiredArgsConstructor
-public class HeatingSystemServiceImpl implements HeatingSystemService {
-    private final HeatingSystemRepository heatingSystemRepository;
-
-    private final KafkaTemplate<String, CommandDto> kafkaTemplate;
-    private final DeviceTypeService deviceTypeService;
-
-    public HeatingSystemServiceImpl(HeatingSystemRepository heatingSystemRepository) {
-        this.heatingSystemRepository = heatingSystemRepository;
-        KafkaProducerConfig kafkaCfg = new KafkaProducerConfig();
-        this.kafkaTemplate = kafkaCfg.kafkaTemplate();
-        this.deviceTypeService = new DeviceTypeService();
-    }
-    
-    @Override
-    public HeatingSystemDto getHeatingSystem(Long id) {
-        HeatingSystem heatingSystem = heatingSystemRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("HeatingSystem not found"));
-        return convertToDto(heatingSystem);
-    }
-
-    @Override
-    public HeatingSystemDto updateHeatingSystem(Long id, HeatingSystemDto heatingSystemDto) {
-        HeatingSystem existingHeatingSystem = heatingSystemRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("HeatingSystem not found"));
-        existingHeatingSystem.setOn(heatingSystemDto.isOn());
-        existingHeatingSystem.setTargetTemperature(heatingSystemDto.getTargetTemperature());
-        HeatingSystem updatedHeatingSystem = heatingSystemRepository.save(existingHeatingSystem);
-        return convertToDto(updatedHeatingSystem);
-    }
-
-    @Override
-    public void turnOn(Long id) {
-        HeatingSystem heatingSystem = heatingSystemRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("HeatingSystem not found"));
-        heatingSystem.setOn(true);
-        heatingSystemRepository.save(heatingSystem);
-
-        this.kafkaTemplate.send("commands", this.deviceTypeService.GetHeatingSystemOnCommand());
-    }
-
-    @Override
-    public void turnOff(Long id) {
-        HeatingSystem heatingSystem = heatingSystemRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("HeatingSystem not found"));
-        heatingSystem.setOn(false);
-        heatingSystemRepository.save(heatingSystem);
-
-        this.kafkaTemplate.send("commands", this.deviceTypeService.GetHeatingSystemOffCommand());
-    }
-
-    @Override
-    public void setTargetTemperature(Long id, double temperature) {
-        HeatingSystem heatingSystem = heatingSystemRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("HeatingSystem not found"));
-        heatingSystem.setTargetTemperature(temperature);
-        heatingSystemRepository.save(heatingSystem);
-
-        this.kafkaTemplate.send("commands", this.deviceTypeService.GetHeatingSystemSetCommand(temperature));
-    }
-
-    @Override
-    public Double getCurrentTemperature(Long id) {
-        // TODO: get from telemetry service
-        HeatingSystem heatingSystem = heatingSystemRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("HeatingSystem not found"));
-        return heatingSystem.getCurrentTemperature();
-    }
-
-    private HeatingSystemDto convertToDto(HeatingSystem heatingSystem) {
-        HeatingSystemDto dto = new HeatingSystemDto();
-        dto.setId(heatingSystem.getId());
-        dto.setOn(heatingSystem.isOn());
-        dto.setTargetTemperature(heatingSystem.getTargetTemperature());
-        return dto;
-    }
-}
