+++ smart-home-monolith/src/main/resources/application.yml
@@ -6,7 +6,10 @@ spring:
     hikari:
       allow-pool-suspension: true
   kafka:
-      bootstrap-servers: ${KAFKA_ADDRESS}
+    bootstrap-servers: ${KAFKA_ADDRESS}
+    producer:
+      key-serializer: org.apache.kafka.common.serialization.StringSerializer
+      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
   jpa:
     hibernate:
       ddl-auto: none
